+++ PowerEditor/src/ScitillaComponent/Buffer.cpp
@@ -1188,7 +1188,7 @@ BufferID FileManager::newEmptyDocument()
 {
 	generic_string newTitle = UNTITLED_STR;
 	TCHAR nb[10];
-	wsprintf(nb, TEXT("%d"), nextUntitledNewNumber());
+	wsprintf(nb, TEXT("%ld"), static_cast<long>(nextUntitledNewNumber()));
 	newTitle += nb;
 
 	Document doc = (Document)_pscratchTilla->execute(SCI_CREATEDOCUMENT);	//this already sets a reference for filemanager
@@ -1205,7 +1205,7 @@ BufferID FileManager::bufferFromDocument(Document doc, bool dontIncrease, bool d
 {
 	generic_string newTitle = UNTITLED_STR;
 	TCHAR nb[10];
-	wsprintf(nb, TEXT("%d"), nextUntitledNewNumber());
+	wsprintf(nb, TEXT("%ld"), static_cast<long>(nextUntitledNewNumber()));
 	newTitle += nb;
 
 	if (!dontRef)
